refactoring convert to process function
