+++ koku/masu/processor/report_processor.py
@@ -17,6 +17,7 @@
 """Report processor external interface."""
 import logging
 
+from django.conf import settings
 from django.db import InterfaceError as DjangoInterfaceError
 from django.db import OperationalError
 from psycopg2 import InterfaceError
@@ -26,9 +27,7 @@ from masu.processor.aws.aws_report_processor import AWSReportProcessor
 from masu.processor.azure.azure_report_processor import AzureReportProcessor
 from masu.processor.gcp.gcp_report_processor import GCPReportProcessor
 from masu.processor.ocp.ocp_report_processor import OCPReportProcessor
-
-# from django.conf import settings
-# from masu.processor.parquet.parquet_report_processor import ParquetReportProcessor
+from masu.processor.parquet.parquet_report_processor import ParquetReportProcessor
 
 LOG = logging.getLogger(__name__)
 
@@ -44,7 +43,7 @@ class ReportProcessorDBError(Exception):
 class ReportProcessor:
     """Interface for masu to use to processor CUR."""
 
-    def __init__(self, schema_name, report_path, compression, provider, provider_uuid, manifest_id):
+    def __init__(self, schema_name, report_path, compression, provider, provider_uuid, manifest_id, context=None):
         """Set the processor based on the data provider."""
         self.schema_name = schema_name
         self.report_path = report_path
@@ -52,6 +51,7 @@ class ReportProcessor:
         self.provider_type = provider
         self.provider_uuid = provider_uuid
         self.manifest_id = manifest_id
+        self.context = context
         try:
             self._processor = self._set_processor()
         except Exception as err:
@@ -73,14 +73,16 @@ class ReportProcessor:
             (Object) : Provider-specific report processor
 
         """
-        # if settings.ENABLE_PARQUET_PROCESSING:
-        #    return ParquetReportProcessor(
-        #        schema_name=self.schema_name,
-        #        report_path=self.report_path,
-        #        compression=self.compression,
-        #        provider_uuid=self.provider_uuid,
-        #        manifest_id=self.manifest_id,
-        #    )
+        if settings.ENABLE_PARQUET_PROCESSING:
+            return ParquetReportProcessor(
+                schema_name=self.schema_name,
+                report_path=self.report_path,
+                compression=self.compression,
+                provider_uuid=self.provider_uuid,
+                provider_type=self.provider_type,
+                manifest_id=self.manifest_id,
+                context=self.context,
+            )
         if self.provider_type in (Provider.PROVIDER_AWS, Provider.PROVIDER_AWS_LOCAL):
             return AWSReportProcessor(
                 schema_name=self.schema_name,
