+++ koku/masu/test/processor/parquet/__init__.py
@@ -1,0 +1,1 @@
+# noqa
