+++ koku/masu/test/processor/test_tasks.py
@@ -55,7 +55,6 @@ from masu.processor._tasks.process import _process_report_file
 from masu.processor.expired_data_remover import ExpiredDataRemover
 from masu.processor.report_processor import ReportProcessorError
 from masu.processor.tasks import autovacuum_tune_schema
-from masu.processor.tasks import convert_to_parquet
 from masu.processor.tasks import get_report_files
 from masu.processor.tasks import record_all_manifest_files
 from masu.processor.tasks import record_report_status
@@ -427,101 +426,6 @@ class TestProcessorTasks(MasuTestCase):
 
         get_report_files(**self.get_report_args)
         mock_cache_remove.assert_called()
-
-    def test_convert_to_parquet(self):
-        """Test the convert_to_parquet task."""
-        logging.disable(logging.NOTSET)
-        expected_logs = [
-            "missing required argument: request_id",
-            "missing required argument: account",
-            "missing required argument: provider_uuid",
-        ]
-        with self.assertLogs("masu.processor.tasks", level="INFO") as logger:
-            with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-                convert_to_parquet(None, None, None, None, "start_date", "manifest_id", [])
-                for expected in expected_logs:
-                    self.assertIn(expected, " ".join(logger.output))
-
-        expected = "Skipping convert_to_parquet. Parquet processing is disabled."
-        with self.assertLogs("masu.processor.tasks", level="INFO") as logger:
-            convert_to_parquet(
-                "request_id", "account", "provider_uuid", "provider_type", "start_date", "manifest_id", "csv_file"
-            )
-            self.assertIn(expected, " ".join(logger.output))
-
-        expected = "Parquet processing is enabled, but no start_date was given for processing."
-        with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-            with self.assertLogs("masu.processor.tasks", level="INFO") as logger:
-                convert_to_parquet(
-                    "request_id", "account", "provider_uuid", "provider_type", None, "manifest_id", "csv_file"
-                )
-                self.assertIn(expected, " ".join(logger.output))
-
-        expected = "Parquet processing is enabled, but the start_date was not a valid date string ISO 8601 format."
-        with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-            with self.assertLogs("masu.processor.tasks", level="INFO") as logger:
-                convert_to_parquet(
-                    "request_id", "account", "provider_uuid", "provider_type", "bad_date", "manifest_id", "csv_file"
-                )
-                self.assertIn(expected, " ".join(logger.output))
-
-        with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.processor.tasks.get_path_prefix"):
-                with patch("masu.processor.tasks.get_file_keys_from_s3_with_manifest_id", return_value=["cur.csv.gz"]):
-                    with patch("masu.processor.tasks.remove_files_not_in_set_from_s3_bucket"):
-                        with patch("masu.processor.tasks.convert_csv_to_parquet"):
-                            convert_to_parquet(
-                                "request_id",
-                                "account",
-                                "provider_uuid",
-                                "AWS",
-                                "2020-01-01T12:00:00",
-                                "manifest_id",
-                                "csv_file",
-                            )
-
-        expected = "Failed to convert the following files to parquet"
-        with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.processor.tasks.get_path_prefix"):
-                with patch(
-                    "masu.processor.tasks.get_file_keys_from_s3_with_manifest_id", return_value=["cost_export.csv"]
-                ):
-                    with patch("masu.processor.tasks.convert_csv_to_parquet", return_value=False):
-                        with self.assertLogs("masu.processor.tasks", level="INFO") as logger:
-                            convert_to_parquet(
-                                "request_id",
-                                "account",
-                                "provider_uuid",
-                                "provider_type",
-                                "2020-01-01T12:00:00",
-                                "manifest_id",
-                                "csv_file",
-                            )
-                            self.assertIn(expected, " ".join(logger.output))
-
-        with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.processor.tasks.get_path_prefix"):
-                with patch(
-                    "masu.processor.tasks.get_file_keys_from_s3_with_manifest_id", return_value=["storage_usage.csv"]
-                ):
-                    with patch("masu.processor.tasks.convert_csv_to_parquet"):
-                        convert_to_parquet(
-                            "request_id",
-                            "account",
-                            "provider_uuid",
-                            "OCP",
-                            "2020-01-01T12:00:00",
-                            "manifest_id",
-                            "csv_file",
-                        )
-
-        with patch("masu.processor.tasks.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.processor.tasks.get_path_prefix"):
-                with patch("masu.processor.tasks.get_file_keys_from_s3_with_manifest_id", return_value=[]):
-                    with patch("masu.processor.tasks.convert_csv_to_parquet"):
-                        convert_to_parquet(
-                            "request_id", "account", "provider_uuid", "OCP", "2020-01-01T12:00:00", "manifest_id"
-                        )
 
 
 class TestRemoveExpiredDataTasks(MasuTestCase):
